Fix #34, used object_parser to cleanup non ingredient components of recipe objects
